feat: Refactor Bookmark components to use new UI elements and improve styling
