+++ client/src/components/ui/Button.tsx
@@ -15,6 +15,7 @@ const buttonVariants = cva(
         secondary: 'bg-secondary text-secondary-foreground hover:bg-secondary/80',
         ghost: 'hover:bg-accent hover:text-accent-foreground',
         link: 'text-primary underline-offset-4 hover:underline',
+        submit: 'bg-surface-submit text-text-primary hover:bg-surface-submit/80',
       },
       size: {
         default: 'h-10 px-4 py-2',
